make ExtendRetractInternal private to ExtendRetract so that no other class can instantiate this class. this ensures that whenever ExtendRetract is called, it always first calls initialize...
