+++ src/main/java/frc/robot/commands/groupCommands/samArmRotExtRetCommands/MiddleScorePosition.java
@@ -6,7 +6,6 @@ package frc.robot.commands.groupCommands.samArmRotExtRetCommands;
 
 import frc.robot.Robot;
 import frc.robot.commands.samArmExtendRetractCommands.ExtendRetract;
-import frc.robot.commands.samArmExtendRetractCommands.ExtendRetractInternal;
 import frc.robot.commands.samArmRotateCommands.RotateArm;
 import frc.robot.rewrittenWPIclasses.SequentialCommandGroup4905;
 import frc.robot.subsystems.samArmExtRet.SamArmExtRetBase;
@@ -27,13 +26,6 @@ public class MiddleScorePosition extends SequentialCommandGroup4905 {
   private boolean m_cube = false;
   private boolean m_backwards = false;
 
-  public MiddleScorePosition(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet) {
-    addCommands(
-        new RotateArm(armRotate, ArmRotationExtensionSingleton.getInstance().getAngle(), true),
-        new ExtendRetractInternal(armExtRet,
-            ArmRotationExtensionSingleton.getInstance().getPosition(), true));
-  }
-
   public MiddleScorePosition(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet, boolean auto,
       boolean cube, boolean backwards) {
     m_auto = auto;
@@ -44,12 +36,16 @@ public class MiddleScorePosition extends SequentialCommandGroup4905 {
         new ExtendRetract(armExtRet));
   }
 
+  public MiddleScorePosition(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet) {
+    // cube and backwards booleans are ignored when not in auto
+    this(armRotate, armExtRet, false, true, true);
+  }
+
   // Called when the command is initially scheduled.
   @Override
   public void additionalInitialize() {
     if (m_auto) {
       if (m_cube) {
-
         if (m_backwards) {
           ArmRotationExtensionSingleton.getInstance().setAngle(m_cubeBackwardMiddleAngle);
           ArmRotationExtensionSingleton.getInstance().setPosition(m_cubeBackwardMiddlePosition);
