+++ src/main/java/frc/robot/commands/groupCommands/samArmRotExtRetCommands/TopScorePosition.java
@@ -5,7 +5,7 @@
 package frc.robot.commands.groupCommands.samArmRotExtRetCommands;
 
 import frc.robot.Robot;
-import frc.robot.commands.samArmExtendRetractCommands.ExtendRetractInternal;
+import frc.robot.commands.samArmExtendRetractCommands.ExtendRetract;
 import frc.robot.commands.samArmRotateCommands.RotateArm;
 import frc.robot.rewrittenWPIclasses.SequentialCommandGroup4905;
 import frc.robot.subsystems.samArmExtRet.SamArmExtRetBase;
@@ -26,13 +26,6 @@ public class TopScorePosition extends SequentialCommandGroup4905 {
   private boolean m_cube = false;
   private boolean m_backwards = false;
 
-  public TopScorePosition(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet) {
-    addCommands(
-        new RotateArm(armRotate, ArmRotationExtensionSingleton.getInstance().getAngle(), true),
-        new ExtendRetractInternal(armExtRet,
-            ArmRotationExtensionSingleton.getInstance().getPosition(), true));
-  }
-
   public TopScorePosition(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet, boolean auto,
       boolean cube, boolean backwards) {
     m_auto = auto;
@@ -40,8 +33,12 @@ public class TopScorePosition extends SequentialCommandGroup4905 {
     m_backwards = backwards;
     addCommands(
         new RotateArm(armRotate, ArmRotationExtensionSingleton.getInstance().getAngle(), true),
-        new ExtendRetractInternal(armExtRet,
-            ArmRotationExtensionSingleton.getInstance().getPosition(), true));
+        new ExtendRetract(armExtRet));
+  }
+
+  public TopScorePosition(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet) {
+    // cube and backwards booleans are ignored when not in auto
+    this(armRotate, armExtRet, false, true, true);
   }
 
   // Called when the command is initially scheduled.
