+++ src/main/java/frc/robot/oi/SmartDashboard4905.java
@@ -31,7 +31,7 @@ import frc.robot.commands.groupCommands.samArmRotExtRetCommands.StowPosition;
 import frc.robot.commands.groupCommands.samArmRotExtRetCommands.SubstationPickupPosition;
 import frc.robot.commands.groupCommands.samArmRotExtRetCommands.TopScorePosition;
 import frc.robot.commands.limeLightCommands.ToggleLimelightLED;
-import frc.robot.commands.samArmExtendRetractCommands.ExtendRetractInternal;
+import frc.robot.commands.samArmExtendRetractCommands.ExtendRetract;
 import frc.robot.commands.samArmRotateCommands.EnableArmBrake;
 import frc.robot.commands.samArmRotateCommands.RotateArm;
 import frc.robot.commands.showBotCannon.PressurizeCannon;
@@ -98,8 +98,8 @@ public class SmartDashboard4905 {
       SmartDashboard.putData("Enable Arm Rotation Brake",
           new EnableArmBrake(subsystemsContainer.getArmRotateBase()));
       SmartDashboard.putNumber("Extend Arm Position Value", 0);
-      SmartDashboard.putData("Extend Arms", new ExtendRetractInternal(
-          subsystemsContainer.getArmExtRetBase(), () -> 5, true, true, false));
+      SmartDashboard.putData("Extend Arms Tuner",
+          new ExtendRetract(subsystemsContainer.getArmExtRetBase(), true, true));
       SmartDashboard.putData("Arm Angle Tuner",
           new RotateArm(subsystemsContainer.getArmRotateBase(), () -> 0, false, true));
     }
