+++ src/main/java/frc/robot/subsystems/SubsystemsContainer.java
@@ -11,7 +11,7 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Config4905;
 import frc.robot.commands.SAMgripperCommands.DefaultGripperCommand;
 import frc.robot.commands.driveTrainCommands.TeleOpCommand;
-import frc.robot.commands.samArmExtendRetractCommands.ExtendRetractInternal;
+import frc.robot.commands.samArmExtendRetractCommands.ExtendRetract;
 import frc.robot.commands.samArmExtendRetractCommands.InitializeArmExtRet;
 import frc.robot.commands.samArmRotateCommands.EnableArmBrake;
 import frc.robot.commands.showBotCannon.AdjustElevation;
@@ -256,7 +256,7 @@ public class SubsystemsContainer {
     }
     if (Config4905.getConfig4905().doesSamArmExtRetExist()) {
       m_armExtRet.setDefaultCommand(new SequentialCommandGroup(new InitializeArmExtRet(m_armExtRet),
-          new ExtendRetractInternal(m_armExtRet, false)));
+          new ExtendRetract(m_armExtRet, false)));
     }
     if (Config4905.getConfig4905().doesSamArmRotateExist()) {
       m_armRotate.setDefaultCommand(new EnableArmBrake(m_armRotate));
